Moved components to updated locations frontend
