+++ frontEnd/src/app/core/services/auth/auth.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { ApiService } from './api.service';
+import { AuthService } from './auth.service';
 
-describe('ApiService', () => {
-  let service: ApiService;
+describe('AuthService', () => {
+  let service: AuthService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(ApiService);
+    service = TestBed.inject(AuthService);
   });
 
   it('should be created', () => {
