--- frontEnd/src/app/login/login.component.html
@@ -1,45 +1,0 @@
-<html lang="en">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <title>Login</title>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-aFq/bzH65dt+w6FI2ooMVUpc+21e0SRygnTpmBvdBgSdnuTN7QbdgL+OapgHtvPp" crossorigin="anonymous">
-    </head>
-    <body>
-        <div class="container">
-            <div class="row justify-content-center">
-                <div class="col-md-6">
-                    <div class="card px-5 py-5">
-                        <div class="form-data">
-                            <form #submitForm="ngForm" (submit)="loginUser(submitForm.value)">
-                                <div class="forms-inputs mb-4">
-                                    <span>Email</span>
-                                    <input type="email" required ngModel #email="ngModel" name="email" placeholder="Enter Email"/>
-                                    <div [hidden]="email.valid || email.pristine" class="alert alert-danger">
-                                        Email is required
-                                    </div>
-                                </div>
-                                
-                                <div class="forms-inputs mb-4">
-                                    <span>Password</span>
-                                    <input type="password" required ngModel #password="ngModel" name="password" placeholder="Enter Password" />
-                                    <div [hidden]="password.valid || password.pristine" class="alert alert-danger">
-                                        Password is required
-                                    </div>
-                                </div>
-
-                                <div class="mb-3">
-                                    <button class="btn btn-dark w-100" [disabled]="!submitForm.form.valid">Login</button>
-                                </div>
-
-                                <a routerLink="/register">Click here to register user</a> 
-                            </form>  
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </body>
-</html>
-
-
