+++ frontEnd/src/app/modules/auth/login/login.component.html
@@ -1,0 +1,37 @@
+<body>
+    <div class="container">
+        <div class="row justify-content-center">
+            <div class="col-md-6">
+                <div class="card px-5 py-5">
+                    <div class="form-data">
+                        <form #submitForm="ngForm" (submit)="loginUser(submitForm.value)">
+                            <div class="forms-inputs mb-4">
+                                <span>Email</span>
+                                <input type="email" required ngModel #email="ngModel" name="email" placeholder="Enter Email"/>
+                                <div [hidden]="email.valid || email.pristine" class="alert alert-danger">
+                                    Email is required
+                                </div>
+                            </div>
+                            
+                            <div class="forms-inputs mb-4">
+                                <span>Password</span>
+                                <input type="password" required ngModel #password="ngModel" name="password" placeholder="Enter Password" />
+                                <div [hidden]="password.valid || password.pristine" class="alert alert-danger">
+                                    Password is required
+                                </div>
+                            </div>
+
+                            <div class="mb-3">
+                                <button class="btn btn-dark w-100" [disabled]="!submitForm.form.valid">Login</button>
+                            </div>
+
+                            <a routerLink="/register">Click here to register user</a> 
+                        </form>  
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</body>
+
+
