+++ frontEnd/src/app/modules/auth/register/register.component.css
